aria/apps/tts-service/requirements.txt
coqui-tts==0.24.3
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.0
soundfile==0.13.1
numpy==1.26.4
# Index: https://download.pytorch.org/whl/cu124
